requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cssselect2==0.8.0
Django==4.2.23
django-celery-beat==2.8.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
exceptiongroup==1.3.0
fonttools==4.58.4
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.4
model-bakery==1.20.5
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.9.0
pyphen==0.17.2
pytest==8.4.1
pytest-django==4.11.1
pytest-cov
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
reportlab==4.4.2
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
coverage
model_bakery